# smx-admin/backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "refresh_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Admin login with email and password",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout and revoke all refresh tokens",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/search-keys/user/{userId}": {
      "get": {
        "operationId": "SearchKeysController_getKeysByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns user's search keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchKey"
                  }
                },
                "examples": {
                  "Example Response": {
                    "summary": "Example user search keys response",
                    "value": [
                      {
                        "id": "123e4567-e89b-12d3-a456-426614174000",
                        "key": "abc123def456",
                        "status": "AVAILABLE",
                        "expiresAt": "2024-12-31T23:59:59Z",
                        "isUsed": false,
                        "usedAt": null,
                        "price": 10,
                        "discountApplied": 5,
                        "stripePaymentId": "pi_1234567890",
                        "createdAt": "2024-01-01T00:00:00Z",
                        "updatedAt": "2024-01-01T00:00:00Z"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "summary": "Get all search keys for a specific user",
        "tags": [
          "search-keys"
        ]
      }
    },
    "/search-keys/cancel/{keyId}": {
      "post": {
        "operationId": "SearchKeysController_cancelKey",
        "parameters": [
          {
            "name": "keyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search key cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchKey"
                },
                "examples": {
                  "Example Response": {
                    "summary": "Example cancelled search key response",
                    "value": {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "key": "abc123def456",
                      "status": "CANCELLED",
                      "expiresAt": "2024-12-31T23:59:59Z",
                      "isUsed": false,
                      "usedAt": null,
                      "price": 10,
                      "discountApplied": 5,
                      "stripePaymentId": "pi_1234567890",
                      "createdAt": "2024-01-01T00:00:00Z",
                      "updatedAt": "2024-01-01T00:00:00Z"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Cancel a search key",
        "tags": [
          "search-keys"
        ]
      }
    },
...
    "/search-keys/payment-details/{stripePaymentId}": {
      "get": {
        "operationId": "SearchKeysController_getPaymentDetails",
        "parameters": [
          {
            "name": "stripePaymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns payment details by stripePaymentId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/"
                },
                "examples": {
                  "Example Response": {
                    "summary": "Example payment details response",
                    "value": {
                      "paymentStatus": "succeeded",
                      "paymentAmount": 100,
                      "paymentCurrency": "USD",
                      "paymentReceiptUrl": "https://example.com/receipt",
                      "paymentCreated": 1718851200
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get payment details by stripePaymentId",
        "tags": [
          "search-keys"
        ]
      }
    }
  },
  "info": {
    "title": "Admin Dashboard API",
    "description": "The Admin Dashboard API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email address of the admin",
            "example": "admin@example.com"
          },
          "password": {
            "type": "string",
            "description": "The password for the admin account",
            "example": "Admin123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "SearchKey": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}